Add basic support for tag-filtered rss feed.
Start a line with "::" in an entry, and it's a list of tags.  Example:
	::tag1,tag2,tag3

Then you can get a filtered rss feed by loading rss.php?tags=foo,a,b.  If no
filter is supplied, we include all entries.  Otherwise, we include an entry
if *any* of its tags match *any* of the requested tags.